+++ examples/clickhouse_integrations/kafka_source/project.tf
@@ -1,0 +1,4 @@
+// Sample project name
+resource "aiven_project" "clickhouse_kafka_source" {
+  project = "clickhouse-kafka-source"
+}